+++ assets/js/script.js
@@ -1,5 +1,5 @@
 (function($) {
-"use strict";
+'use strict';
 
 /* --------------
   GLOBAL FORM
@@ -14,7 +14,7 @@ GlobalForm.prototype = {
   addListener: function() {
     var that = this;
     var globalVal = 0;
-    $("#variable_product_options").on("change", GLOBAL[that.field], _onChange);
+    $('#variable_product_options').on('change', GLOBAL[that.field], _onChange);
 
     /*
       Listener
@@ -24,7 +24,7 @@ GlobalForm.prototype = {
 
       // change placeholder on quick field
       var placeholder = (globalVal) ? globalVal : that.defaultPlaceholder;
-      $(QUICK[that.field]).attr("placeholder", placeholder);
+      $(QUICK[that.field]).attr('placeholder', placeholder);
 
       $(MAIN[that.field] ).val(globalVal); // change main field
       $(TARGET[that.field] ).each(_applyToVariation); // change all variations field
@@ -32,7 +32,7 @@ GlobalForm.prototype = {
 
     function _applyToVariation() {
       var $target = $(this);
-      var $quick = $(this).closest(".woocommerce_variation").find(QUICK[that.field] );
+      var $quick = $(this).closest('.woocommerce_variation').find(QUICK[that.field] );
 
       var isEmpty = !$quick.val();
       var isEqualGlobal = $quick.val() === globalVal;
@@ -41,7 +41,7 @@ GlobalForm.prototype = {
       if(isEmpty) { $target.val(globalVal).change(); }
 
       // if equal global, remove the content so it shows only the placeholder
-      if(isEqualGlobal) { $quick.val(""); }
+      if(isEqualGlobal) { $quick.val(''); }
     }
   },
 };
@@ -52,9 +52,9 @@ GlobalForm.create = function() {
 
   var data = { currency: woocommerce_admin_meta_boxes.currency_format_symbol };
 
-  var template = Handlebars.compile($("#h-global-form").html() );
+  var template = Handlebars.compile($('#h-global-form').html() );
   var html = template(data);
-  $("#variable_product_options .toolbar-top").after(html);
+  $('#variable_product_options .toolbar-top').after(html);
 
   // copy the number from main form to global form
   $(GLOBAL.price).val( $(MAIN.price).val() );
@@ -72,25 +72,29 @@ function QuickForm(args) {
 QuickForm.prototype = {
   addListener: function() {
     var that = this;
-    $("#variable_product_options").on("change", QUICK[that.field], _onChange);
+    $('#variable_product_options').on('change', QUICK[that.field], _onChange);
 
     /*
       Listener
     */
     function _onChange(e) {
-      var quickVal = $(this).val();
+      var newVal = $(this).val();
+      var $target = $(this).closest('.woocommerce_variation').find(TARGET[that.field]);
 
-      // change the target field to be the same
-      var $target = $(this).closest(".woocommerce_variation").find(TARGET[that.field]);
-      $target.val(quickVal);
-
-      if(that.field === "price" || that.field === "sale") {
-        // if same as global value, empty out the quick field so it shows the placeholder
+      if(that.field === 'price' || that.field === 'sale') {
+        // if same as globalVal, empty out the quick field so it shows the placeholder
         var globalVal = $(GLOBAL[that.field] ).val();
-        if(quickVal === globalVal) {
-          $(this).val("");
+        if(newVal === globalVal) {
+          $(this).val('');
+        }
+        // if empty but global val not empty, assign the globalVal to target
+        else if(!newVal && globalVal) {
+          newVal = globalVal;
         }
       }
+
+      // change the target field to be the same
+      $target.val(newVal);
     } // _onChange
 
   },
@@ -102,21 +106,21 @@ QuickForm.prototype = {
   @param int $variation - The outer container of the variation
 */
 QuickForm.create = function($variation) {
-  var $header = $variation.closest(".woocommerce_variation").find("h3");
+  var $header = $variation.closest('.woocommerce_variation').find('h3');
 
   var data = _parseData($variation);
 
   // manage stock always on
-  $variation.find(TARGET.manageStock).prop("checked", true).change();
+  $variation.find(TARGET.manageStock).prop('checked', true).change();
 
   // append template
-  var template = Handlebars.compile($("#h-quick-form").html() );
+  var template = Handlebars.compile($('#h-quick-form').html() );
   var html = template(data);
   $header.append(html);
 
   // trigger change in target field
-  $variation.find(TARGET.price).val(data._regular_price).change();
-  $variation.find(TARGET.sale).val(data._sale_price).change();
+  // $variation.find(TARGET.price).val(data._regular_price).change();
+  // $variation.find(TARGET.sale).val(data._sale_price).change();
 
   /*
     Get the raw data and process it
@@ -125,13 +129,17 @@ QuickForm.create = function($variation) {
   */
   function _parseData() {
     // parse data
-    var data = $variation.find(".h-variation-data").data("variation");
+    var data = $variation.find('.h-variation-data').data('variation');
 
     data.currency = woocommerce_admin_meta_boxes.currency_format_symbol;
     data.globalPrice = $(GLOBAL.price).val();
     data.globalSale = $(GLOBAL.sale).val();
-    data.isEqualGlobalPrice = data.globalPrice === data._regular_price;
-    data.isEqualGlobalSale = data.globalSale === data._sale_price;
+
+    data._regular_price = parseInt(data._regular_price);
+    data._sale_price = parseInt(data._sale_price);
+
+    data.isEqualGlobalPrice = data.globalPrice == data._regular_price;
+    data.isEqualGlobalSale = data.globalSale == data._sale_price;
 
     // if price empty
     if(!data._regular_price) {
@@ -160,7 +168,7 @@ function TargetForm(args) {
 TargetForm.prototype = {
   addListener: function() {
     var that = this;
-    $("#variable_product_options").on("change", TARGET[that.field], _onChange);
+    $('#variable_product_options').on('change', TARGET[that.field], _onChange);
 
     /*
       Listener
@@ -169,11 +177,11 @@ TargetForm.prototype = {
       var targetVal = $(this).val();
 
       // change the quick field to be the same
-      var $quick = $(this).closest(".woocommerce_variation").find(QUICK[that.field] );
+      var $quick = $(this).closest('.woocommerce_variation').find(QUICK[that.field] );
       $quick.val(targetVal);
 
       switch(that.field) {
-        case "price":
+        case 'price':
           // if empty, set it to 0
           if(!targetVal) {
             targetVal = 0;
@@ -181,23 +189,23 @@ TargetForm.prototype = {
           }
           break;
 
-        case "sale":
-          var $form = $quick.closest(".quick-form");
+        case 'sale':
+          var $form = $quick.closest('.quick-form');
 
           // if empty, remove the has-sale class
           if(!targetVal) {
-            $form.removeClass("has-sale");
+            $form.removeClass('has-sale');
           } else {
-            $form.addClass("has-sale");
+            $form.addClass('has-sale');
           }
           break;
       }
 
-      if(that.field === "price" || that.field === "sale") {
+      if(that.field === 'price' || that.field === 'sale') {
         // if same as global value, empty out the quick field so it shows the placeholder
         var globalVal = $(GLOBAL[that.field]).val();
         if(targetVal === globalVal) {
-          $quick.val("");
+          $quick.val('');
         }
       }
     }  // _onChange
@@ -218,17 +226,17 @@ var changeDefault = {
     var that = this;
 
     // remove default toggle
-    $("#variable_product_options").off("click", ".wc-metabox h3");
-    $("#variable_product_options").off("click", ".wc-metabox > h3");
+    $('#variable_product_options').off('click', '.wc-metabox h3');
+    $('#variable_product_options').off('click', '.wc-metabox > h3');
 
     // toggle only if clicking the Edit button
-    $("#variable_product_options").on("click", "h3 .handlediv", that.toggleVariation);
+    $('#variable_product_options').on('click', 'h3 .handlediv', that.toggleVariation);
   },
 
   toggleVariation: function() {
-    var $container = $(this).closest(".woocommerce_variation");
-    $container.toggleClass("open closed");
-    $container.find(".wc-metabox-content").slideToggle();
+    var $container = $(this).closest('.woocommerce_variation');
+    $container.toggleClass('open closed');
+    $container.find('.wc-metabox-content').slideToggle();
   }
 };
 
@@ -237,62 +245,68 @@ var changeDefault = {
 ----------- */
 
 var TARGET = {
-  form: ".woocommerce_variation",
-  price: "[name*='variable_regular_price']",
-  stock: "[name*='variable_stock[']",
-  manageStock: "[name*='variable_manage_stock']",
-  sale: "[name*='variable_sale_price[']",
+  form: '.woocommerce_variation',
+  price: '[name*="variable_regular_price"]',
+  stock: '[name*="variable_stock["]',
+  manageStock: '[name*="variable_manage_stock"]',
+  sale: '[name*="variable_sale_price["]',
 };
 
 var GLOBAL = {
-  form: ".global-form",
-  price: "#global-price",
-  sale: "#global-sale",
+  form: '.global-form',
+  price: '#global-price',
+  sale: '#global-sale',
 };
 
 var QUICK = {
-  form: ".quick-form",
-  price: "[name*='quick_price']",
-  sale: "[name*='quick_sale']",
-  stock: "[name*='quick_stock']",
+  form: '.quick-form',
+  price: '[name*="quick_price"]',
+  sale: '[name*="quick_sale"]',
+  stock: '[name*="quick_stock"]',
 };
 
 var MAIN = {
-  price: "#_regular_price",
-  sale: "#_sale_price",
+  price: '#_regular_price',
+  sale: '#_sale_price',
 };
+
+var BACKORDER = {
+  main: '#_backorders',
+  target: '[name*="variable_backorders["]'
+}
 
 /* -----------------
   VARIATION Handler
 ----------------- */
 var variation = {
   init: function() {
-    var $variationTab = $("#variable_product_options");
+    var $variationTab = $('#variable_product_options');
 
-    // $variation.on("DOMNodeInserted", this.onAdded);
-    $("#woocommerce-product-data").on("woocommerce_variations_loaded", this.onLoaded);
-    $("#woocommerce-product-data").on("woocommerce_variations_added", this.onAdded);
-    $("#woocommerce-product-data").on("woocommerce_variations_saved", this.onSaved);
+    GlobalForm.create();
 
-    $variationTab.on("focus", "h3 input", this.highlightHeader);
-    $variationTab.on("blur", "h3 input", this.removeHighlightHeader);
+    // $variation.on('DOMNodeInserted', this.onAdded);
+    $('#woocommerce-product-data').on('woocommerce_variations_loaded', this.onLoaded);
+    $('#woocommerce-product-data').on('woocommerce_variations_added', this.onAdded);
+    $('#woocommerce-product-data').on('woocommerce_variations_saved', this.onSaved);
+
+    $variationTab.on('focus', 'h3 input', this.highlightHeader);
+    $variationTab.on('blur', 'h3 input', this.removeHighlightHeader);
 
     this.addListener();
   },
 
   highlightHeader: function(e) {
-    $(this).closest(".woocommerce_variation").addClass("active");
+    $(this).closest('.woocommerce_variation').addClass('active');
   },
 
   removeHighlightHeader: function(e) {
-    $(this).closest(".woocommerce_variation").removeClass("active");
+    $(this).closest('.woocommerce_variation').removeClass('active');
   },
 
   /*
     After finished loading the Variations tab
   */
   onLoaded: function(e) {
-    console.log("onLoaded");
     var that = variation;
 
     // if no global form yet
@@ -300,7 +314,7 @@ var variation = {
       GlobalForm.create();
     }
 
-    $(".woocommerce_variation").each(function() {
+    $('.woocommerce_variation').each(function() {
       QuickForm.create($(this) );
     });
   },
@@ -309,12 +323,11 @@ var variation = {
     After adding a new variation using the toolbar
   */
   onAdded: function(e) {
-    console.log("onAdded");
     var that = variation;
-    var $newVar = $(this).find(".woocommerce_variation:not(:has(.quick-form))");
+    var $newVar = $(this).find('.woocommerce_variation:not(:has(.quick-form))');
 
     // if adding one variation (if multiple, it won't have the class below)
-    if($newVar.is(".variation-needs-update") ) {
+    if($newVar.is('.variation-needs-update') ) {
       QuickForm.create($newVar);
     }
   },
@@ -323,10 +336,9 @@ var variation = {
     After AJAX Save the variatios
   */
   onSaved: function(e) {
-    console.log("onSaved");
 
     var post = {
-      action: "h_after_save_variations",
+      action: 'h_after_save_variations',
       data: {
         post_id: woocommerce_admin_meta_boxes.post_id,
         global_price: $(GLOBAL.price).val(),
@@ -347,17 +359,17 @@ var variation = {
   },
 
   globalListener: function() {
-    var priceField = new GlobalForm({ field: "price", defaultPlaceholder: "Price" });
-    var saleField = new GlobalForm({ field: "sale", defaultPlaceholder: "Sale" });
+    var priceField = new GlobalForm({ field: 'price', defaultPlaceholder: 'Price' });
+    var saleField = new GlobalForm({ field: 'sale', defaultPlaceholder: 'Sale' });
 
     priceField.addListener();
     saleField.addListener();
   },
 
   targetListener: function() {
-    var priceField = new TargetForm({ field: "price" });
-    var saleField = new TargetForm({ field: "sale" });
-    var stockField = new TargetForm({ field: "stock" });
+    var priceField = new TargetForm({ field: 'price' });
+    var saleField = new TargetForm({ field: 'sale' });
+    var stockField = new TargetForm({ field: 'stock' });
 
     priceField.addListener();
     saleField.addListener();
@@ -365,9 +377,9 @@ var variation = {
   },
 
   quickListener: function() {
-    var priceField = new QuickForm({ field: "price" });
-    var saleField = new QuickForm({ field: "sale" });
-    var stockField = new QuickForm({ field: "stock" });
+    var priceField = new QuickForm({ field: 'price' });
+    var saleField = new QuickForm({ field: 'sale' });
+    var stockField = new QuickForm({ field: 'stock' });
 
     priceField.addListener();
     saleField.addListener();
@@ -381,15 +393,29 @@ var variation = {
 ------------------ */
 var attribute = {
   init: function() {
-    $(document.body).on("woocommerce_added_attribute", this.onAdded);
+    $(document.body).on('woocommerce_added_attribute', this.onAdded);
   },
 
   /*
-    After adding new attribute, auto check the "Used for variations"
+    After adding new attribute, auto check the 'Used for variations'
   */
   onAdded: function(e) {
-    console.log($(this) );
-    $("#product_attributes").find("[name*='attribute_variation[']").prop("checked", true).change();
+    $('#product_attributes').find('[name*="attribute_variation["]').prop('checked', true).change();
+  }
+};
+
+/* ----------------
+  BACKORDER Handler
+---------------- */
+
+var backorder = {
+  init: function() {
+    $(BACKORDER.main).on('change', this.onChange);
+  },
+
+  onChange: function(e) {
+    var newVal = $(this).val();
+    $(BACKORDER.target).val(newVal);
   }
 };
 
@@ -397,19 +423,10 @@ var start = function() {
   changeDefault.init();
   variation.init();
   attribute.init();
+  // backorder.init();
 };
 
 $(document).ready(start);
-$(document).on("page:load", start);
+$(document).on('page:load', start);
 
 })(jQuery);
-
-/* ----------------
-  Public Function
------------------- */
-
-var Hoo = {
-  addQuickForm: function(index, data) {
-    var $ = jQuery;
-  }
-};